* error.c (dump_expr): Fix format specifier warning.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@59867 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/cp/error.c
@@ -1479,7 +1479,8 @@ dump_expr (t, flags)
 	    if (host_integerp (t, TREE_UNSIGNED (type)))
 	      dump_char (tree_low_cst (t, TREE_UNSIGNED (type)));
 	    else
-	      output_printf (scratch_buffer, "\\x%x", TREE_INT_CST_LOW (t));
+	      output_printf (scratch_buffer, "\\x%x",
+			     (unsigned int) TREE_INT_CST_LOW (t));
 	    output_add_character (scratch_buffer, '\'');
 	  }
 	else
